requirements.txt
-e git+git://github.com/Sprtch/despinassy.git#egg=despinassy-pkg-tperale
daemonize
Jinja2
redis
PyYAML
